Avoid default arguments to ScaledIntegerNode since this causes ambiguity. Added E57RefImplConfig.cmake.
git-svn-id: https://svn.code.sf.net/p/e57-3d-imgfmt/code/trunk@146 e8104519-1f18-44c1-aea1-0fd5f35a081b

+++ doc/FoundationAPI/examples/ScaledIntegerCreate.cpp
@@ -34,9 +34,13 @@ int main(int /*argc*/, char** /*argv*/) {
         ImageFile imf("temp._e57", "w");
         StructureNode root = imf.root();
 
-        // Create 5 example ScaledIntegers 
-        root.set("si1", ScaledIntegerNode(imf));
-        root.set("si2", ScaledIntegerNode(imf, 123));
+        // Create 5 example ScaledIntegers
+        // The requirement for the explicit types for the first parameter
+        // are caused by a questionable definition of the ScaledIntegerNode
+        // ctors which pose a ambiguity because all parameters may be optional.
+        // <roland_schwarz, 2011-03-10>
+        root.set("si1", ScaledIntegerNode(imf, int64_t()));
+        root.set("si2", ScaledIntegerNode(imf, int64_t(123)));
         root.set("si3", ScaledIntegerNode(imf, 123, 0, 1023));
         root.set("si4", ScaledIntegerNode(imf, 123, 0, 1023, .001));
         root.set("si5", ScaledIntegerNode(imf, 123, 0, 1023, .001, 100.0));
